requirements.txt
grpcio~=1.70.0
grpcio-tools
pyannote.audio
torch
protobuf~=5.29.3